+++ examples/database/adb/autonomous_database.tf
@@ -28,7 +28,7 @@ resource "oci_database_autonomous_database" "autonomous_database" {
   compartment_id           = var.compartment_ocid
   cpu_core_count           = "1"
   data_storage_size_in_tbs = "1"
-  db_name                  = "adbdb1f"
+  db_name                  = "adbDatabaseName123"
 
   #Optional
   db_version                                     = data.oci_database_autonomous_db_versions.test_autonomous_db_versions.autonomous_db_versions[0].version